src/python/projects/api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
bcrypt==4.0.1
pymongo>=4.0.0
beanie>=1.24.0
celery>=5.3.0
redis>=5.0.0
aiofiles>=23.0.0
httpx>=0.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
motor>=3.3.0
passlib[bcrypt]>=1.7.4
openai>=1.0.0